prevent overwriting generated dataset

+++ openai_generate.py
@@ -236,6 +236,12 @@ def main():
     start_time = time.time()
 
     generation_writer = os.path.join(args.output_dir, "test.tsv")
+
+    # prevent from overwriting generated dataset
+    if os.path.isfile(generation_writer):
+        logger.info('Generated dataset already exists. Exit Program.')
+        exit()
+
     with open(generation_writer, "w") as file_writer:
         tsv_writer = csv.writer(file_writer, delimiter='\t')
         for step, inputs in tqdm(enumerate(eval_dataset)):
